requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.34
pydantic==2.9.0
Jinja2==3.1.4
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
